requirements.txt
python-telegram-bot==20.0
urllib3==1.26.5
